🎉 Versioned Exporting
Allow the user to select a specific Minecraft version to target when exporting.

# Changes
- Created versioned data pack and resource pack compilers.
- Added a Target Minecraft Version blueprint setting.
- Added support for Minecraft 1.20.4.
- Added support for Minecraft 1.20.5.
- Added support for Minecraft 1.21.2.

# Fixes
- Fixed "default is not defined" error when attempting to export a newly converted blueprint.
- Disabled broken zip export modes (will re-enable when fixed).
- Fixed missing default values in locator config UI.
- Fixed bone config types to ensure they're considered optional.
- Updated the packager ESBuild plugin to only copy over prod builds.

+++ tools/plugins/packagerPlugin.ts
@@ -44,8 +44,7 @@ function plugin(): Plugin {
 				if (fs.existsSync(pathjs.join(DIST_PATH, 'about.svelte')))
 					fs.unlinkSync(pathjs.join(DIST_PATH, 'about.svelte'))
 
-				if (fs.existsSync(PLUGIN_REPO_PATH)) {
-					console.log('📋 Copying to plugin repo')
+				if (process.env.NODE_ENV === 'production' && fs.existsSync(PLUGIN_REPO_PATH)) {
 					fs.rmSync(PLUGIN_REPO_PATH, { recursive: true, force: true })
 					fs.cpSync(DIST_PATH, PLUGIN_REPO_PATH, { recursive: true })
 					const manifest = JSON.parse(fs.readFileSync(PLUGIN_MANIFEST_PATH, 'utf-8'))
@@ -57,6 +56,7 @@ function plugin(): Plugin {
 							parser: 'json',
 						})
 					)
+					console.log('📋 Copied to Plugin Repo!')
 				}
 			})
 		},
